requirements.txt
-i https://pypi.douban.com/simple/ --trusted-host=pypi.douban.com/simple
requests==2.20.1
PyExecJs==1.5.1
pycrypto==2.6.1
pkcs7==0.1.2
Pillow==5.3.0
rsa==4.0
matplotlib==3.0.2
flask==1.0.2
flask_mongoengine==0.9.5